requirements.txt
pdfminer.six
transformers
torch
sentencepiece
reportlab
tf-keras
